fix(game): initialize is_winner to false by default
That was made as an attempt to fix a bug stopping the app from compiling.
The main loop was not entering, so I guessed that the is_winner variable
was not being initialized to false by default. Golang is to blame, as the
values always have a default value, even if not explicitly set.