refactor(words_feed): call read_file before picking word
This commit ensures read_file is called before picking a random word to
avoid segmentation faults, which occurs when words list is empty. For
this reason, the selected index is zero, but it stills breaks, because
there is no element at space 0.

Remove unused cstdlib and ctime includes from header file